embed/requirements.txt
fastapi[standard]
uvicorn
model2vec>=0.8.0
numpy
pyyaml
